bin: Handle CLI arguments with while loop
This does not make much difference, since on most branch of case
statement, and `exit` is invoked at the end.

It seems to make a difference in a separate experiment as trialed
outside this project, where the wildcard case will always be reached.

This is a cherry picked commit from archlinux branch.