bin: Encrypt backup files
An overhaul on files backup. Files are now bundled into a single
archive, with encryption in place on the final archive file using AES
256-bit CBC mode. A digest for the encrypted archive is then generated.

tar is used in favour of 7z, since the latter is officially not
recommended for backup purpose.

Option to transfer the archive to external drive is removed, since at
the moment of writing, a flash drive is used as the external drive,
which is not a reliable long-term storage option. I also seldom keep an
external HDD around, so factoring out the option to back up to flash
drive, this option is pretty much going to be unused.