bin: Backup documents and keys separately
Storing documents and keys together in the same archive, and then
storing this archive on a guarded cloud storage platform, will result in
lockout since there wouldn't be a way to access the cloud storage in
case all devices are lost.

Hence, documents and keys are now stored separately. The former will be
encrypted with GPG, and the latter will be encrypted with OpenSSL using
a memorable passphrase, therefore eliminating the use of password
manager in the flow of creating and storing backups.

GPG keys are now backed up by first exporting the keys and ownertrust
first, rather than archiving the entire directory.